+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Version 1.1.0.dev0
 
 - Added support for `HEADER_TEXT` and `FOOTER_TEXT` to go before and after help output
+- Catch Abort exceptions from `cmd+c` and print nicely using `ABORTED_TEXT`
 
 ## Version 1.0.0 (2022-02-18)